feat: moved related functions to avoid error
